+++ lib/presentation/discover_detailed_view_rehab_centers_screen/discover_detailed_view_rehab_centers_screen.dart
@@ -22,7 +22,15 @@ class DiscoverDetailedViewRehabCentersScreen extends StatelessWidget {
     return SafeArea(
         child: Scaffold(
             resizeToAvoidBottomInset: false,
-            appBar: _buildAppBar(context),
+            appBar: AppBar(
+        backgroundColor: Colors.transparent,
+          leading: IconButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            icon: Icon(Icons.arrow_back_ios_new_outlined,),
+          ),
+        ),
             body: Container(
                 width: double.maxFinite,
                 padding: EdgeInsets.symmetric(horizontal: 25.h, vertical: 7.v),
@@ -38,32 +46,10 @@ class DiscoverDetailedViewRehabCentersScreen extends StatelessWidget {
                       style: CustomTextStyles.labelSmallGray600),
                   SizedBox(height: 28.v)
                 ])),
-            bottomNavigationBar: _buildNavigationBar(context),
-            floatingActionButton: CustomFloatingButton(
-                height: 81,
-                width: 81,
-                backgroundColor: appTheme.amber900,
-                child: CustomImageView(
-                    imagePath: ImageConstant.imgLocationPin3,
-                    height: 40.5.v,
-                    width: 40.5.h)),
-            floatingActionButtonLocation:
-                FloatingActionButtonLocation.centerDocked));
+        ),
+      );
   }
 
-  /// Section Widget
-  PreferredSizeWidget _buildAppBar(BuildContext context) {
-    return CustomAppBar(
-        leadingWidth: 52.h,
-        leading: AppbarLeadingIconbuttonOne(
-            imagePath: ImageConstant.imgArrowLeft,
-            margin: EdgeInsets.only(left: 19.h, top: 2.v, bottom: 21.v),
-            onTap: () {
-              onTapArrowLeft(context);
-            }),
-        centerTitle: true,
-        title: AppbarSubtitleOne(text: "All Rehab Centers"));
-  }
 
   /// Section Widget
   Widget _buildUserProfile(BuildContext context) {
@@ -90,12 +76,6 @@ class DiscoverDetailedViewRehabCentersScreen extends StatelessWidget {
             }));
   }
 
-  /// Section Widget
-  Widget _buildNavigationBar(BuildContext context) {
-    return CustomBottomAppBar(onChanged: (BottomBarEnum type) {
-      Navigator.pushNamed(navigatorKey.currentContext!, getCurrentRoute(type));
-    });
-  }
 
   ///Handling route based on bottom click actions
   String getCurrentRoute(BottomBarEnum type) {
